copied some stuff

+++ app/__init__.py
@@ -2,9 +2,9 @@
 # App Creation and Launch
 #===========================================================
 
-from flask import Flask, render_template, request, flash, redirect, session
-import html
+from flask import Flask, render_template, session, request, flash, redirect
 from werkzeug.security import generate_password_hash, check_password_hash
+import html
 
 from app.helpers.session import init_session
 from app.helpers.db import connect_db
@@ -38,6 +38,22 @@ def about():
 
 
 #-----------------------------------------------------------
+# Sign-Up page route
+#-----------------------------------------------------------
+@app.get("/signup/")
+def signup():
+    return render_template("pages/signup.jinja")
+
+
+#-----------------------------------------------------------
+# Login page route
+#-----------------------------------------------------------
+@app.get("/login/")
+def login():
+    return render_template("pages/login.jinja")
+
+
+#-----------------------------------------------------------
 # Things page route - Show all the things, and new thing form
 #-----------------------------------------------------------
 @app.get("/things/")
@@ -59,7 +75,17 @@ def show_all_things():
 def show_one_thing(id):
     with connect_db() as client:
         # Get the thing details from the DB
-        sql = "SELECT id, name, price FROM things WHERE id=?"
+        sql = """
+            SELECT things.id   AS t_id,
+                   things.name AS t_name,
+                   users.name  AS u_name,
+                   users.id    AS u_id
+
+            FROM things
+            JOIN users ON things.user_id = users.id
+
+            WHERE things.id=?
+        """
         values = [id]
         result = client.execute(sql, values)
 
@@ -78,15 +104,20 @@ def show_one_thing(id):
 # Route for adding a thing, using data posted from a form
 #-----------------------------------------------------------
 @app.post("/add")
-def add():
-    # get the data from the form
-    name = request.form.get("name")
-    price = request.form.get("price")
+def add_a_thing():
+    # Get the data from the form
+    name  = request.form.get("name")
+
+    # Sanitise the inputs
+    name = html.escape(name)
+
+    # Get the user_id from the session
+    user_id = session["user_id"]
 
     with connect_db() as client:
-        # Add the things to the DB
-        sql = "INSERT INTO things (name, price) VALUES (?, ?)"
-        values = [name, price]
+        # Add the thing to the DB
+        sql = "INSERT INTO things (name, user_id) VALUES (?, ?)"
+        values = [name, user_id]
         client.execute(sql, values)
 
         # Go back to the home page
@@ -94,36 +125,74 @@ def add():
         return redirect("/things")
 
 
+#-----------------------------------------------------------
+# Route for adding a user, using data posted from a form
+#-----------------------------------------------------------
 @app.post("/add-user")
 def add_a_user():
-    # get the data from the form
-    name = request.form.get("name")
-    username = request.form.get("username")
-    password = request.form.get("password")
+    # Get the data from the form
+    name      = request.form.get("name")
+    username  = request.form.get("username")
+    password  = request.form.get("password")
 
-    #sanatize the inputs
-    name = html.escape
-    username = html.escape
+    # Sanitise the inputs
+    name     = html.escape(name)
+    username = html.escape(username)
 
-    print (name)
-    print (username)
-    print (password)
     # Hash the password
     hash = generate_password_hash(password)
 
-
     with connect_db() as client:
         # Add the user to the DB
-        sql = "INSERT INTO users (name, username, password_hash ) VALUES (?, ?, ?)"
+        sql = "INSERT INTO users (name, username, password_hash) VALUES (?, ?, ?)"
         values = [name, username, hash]
         client.execute(sql, values)
 
-        
+        # Go back to the home page
+        flash(f"Thing '{name}' added", "success")
+        return redirect("/")
+
+
+#-----------------------------------------------------------
+# Route for logging in a user, using data posted from a form
+#-----------------------------------------------------------
+@app.post("/login-user")
+def login_user():
+    # Get the data from the form
+    username  = request.form.get("username")
+    password  = request.form.get("password")
+
+    # Sanitise the inputs
+    username = html.escape(username)
+
+    with connect_db() as client:
+        # Try to find a matching record
+        sql = """
+            SELECT id, name, password_hash
+            FROM users
+            WHERE username=?
+        """
+        values = [username]
+        result = client.execute(sql, values)
+
+        # Check if we got a record
+        if result.rows:
+            # Yes, so user exists
+            user = result.rows[0]
+            hash = user["password_hash"]
+
+            # Check if passwords match
+            if check_password_hash(hash, password):
+                # Yes, so save the details in the session
+                session["user_id"] = user["id"]
+                session["user_name"] = user["name"]
+                flash("Logged in successfully", "success")
+                return redirect("/")
+
 
         # Go back to the home page
-        flash(f"user '{name}' added", "success")
-        return redirect("/")
-    
+        flash("Incorrect credentials", "error")
+        return redirect("/login")
 
 
 #-----------------------------------------------------------
@@ -132,17 +201,26 @@ def add_a_user():
 @app.get("/delete/<int:id>")
 def delete_a_thing(id):
     with connect_db() as client:
+        #get our user id from session
+        user_id = session["user_id"]
         # Delete the thing from the DB
-        sql = "DELETE FROM things WHERE id=?"
-        values = [id]
+        sql = "DELETE FROM things WHERE id=? AND user_id=?"
+        values = [id,user_id]
         client.execute(sql, values)
 
         # Go back to the home page
-        flash("Thing deleted", "warning")
+        flash("Thing deleted", "success")
         return redirect("/things")
 
 
-@app.get("/signup")
-def signup():
-    return render_template("pages/signup.jinja")
-
+#-----------------------------------------------------------
+# Route user logout
+#-----------------------------------------------------------
+@app.get("/logout")
+def logout():
+    # Clear the session values
+    session.pop("user_id")
+    session.pop("user_name")
+    # Back to the home page
+    flash("You have been logged out!", "success")
+    return redirect("/")
